+++ eve-online-intel-tool.php
@@ -76,6 +76,7 @@ class EveOnlineIntelTool {
 
 		if(\is_admin()) {
 			new Libs\Admin\PluginSettings;
+			new Libs\Widgets\Dashboard\CacheStatisticsWidget;
 
 			$this->initGitHubUpdater();
 		} // if(\is_admin())